highscore controller added

+++ database/migrations/2024_06_03_130620_create_highscores_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->timestamps();
             $table->foreignId('deck_id')->constrained('decks')->onDelete('cascade');
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
-            $table->float('time');
+            $table->integer('time');
         });
     }
